refactor(vm): split vsock.rs into a module directory
Pure move, promised on #657 once #658 had taken the handshake and
listener out of the same file. `vsock/mod.rs` keeps the docs, the shared
types (ExitStatus, OutputChunk, ExecInputMsg, PortWait), connect_to_port
with its retry, wait_ready, and the frame codec; each protocol gets its
own file — `exec.rs` (run, exec), `clock.rs` (sync_clock, ClockSync),
`net.rs` (reconfigure_network, ReconfigTimings), `wait_port.rs`
(wait_for_port) — with its tests. No logic changes; the `// ====` section
dividers go with the split that made them redundant, and every
`vsock::…` path callers use is unchanged.
